refactor: layout of login related page
