+++ ui/console-src/router/routes.config.ts
@@ -2,10 +2,13 @@ import type { RouteRecordRaw } from "vue-router";
 import NotFound from "@/views/exceptions/NotFound.vue";
 import Forbidden from "@/views/exceptions/Forbidden.vue";
 import BasicLayout from "@console/layouts/BasicLayout.vue";
+import GatewayLayout from "@console/layouts/GatewayLayout.vue";
 import Setup from "@console/views/system/Setup.vue";
 import Redirect from "@console/views/system/Redirect.vue";
 import SetupInitialData from "@console/views/system/SetupInitialData.vue";
 import ResetPassword from "@console/views/system/ResetPassword.vue";
+import Login from "@console/views/system/Login.vue";
+import Binding from "@console/views/system/Binding.vue";
 
 export const routes: Array<RouteRecordRaw> = [
   {
@@ -25,12 +28,46 @@ export const routes: Array<RouteRecordRaw> = [
     ],
   },
   {
+    path: "/login",
+    component: GatewayLayout,
+    children: [
+      {
+        path: "",
+        name: "Login",
+        component: Login,
+        meta: {
+          title: "core.login.title",
+        },
+      },
+    ],
+  },
+  {
+    path: "/binding/:provider",
+    component: GatewayLayout,
+    children: [
+      {
+        path: "",
+        name: "Binding",
+        component: Binding,
+        meta: {
+          title: "core.binding.title",
+        },
+      },
+    ],
+  },
+  {
     path: "/setup",
-    name: "Setup",
-    component: Setup,
-    meta: {
-      title: "core.setup.title",
-    },
+    component: GatewayLayout,
+    children: [
+      {
+        path: "",
+        name: "Setup",
+        component: Setup,
+        meta: {
+          title: "core.setup.title",
+        },
+      },
+    ],
   },
   {
     path: "/setup-initial-data",
@@ -47,11 +84,17 @@ export const routes: Array<RouteRecordRaw> = [
   },
   {
     path: "/reset-password",
-    name: "ResetPassword",
-    component: ResetPassword,
-    meta: {
-      title: "core.reset_password.title",
-    },
+    component: GatewayLayout,
+    children: [
+      {
+        path: "",
+        name: "ResetPassword",
+        component: ResetPassword,
+        meta: {
+          title: "core.reset_password.title",
+        },
+      },
+    ],
   },
 ];
